textproc/utf8proc: finish disabling test
Commit 34693bdefa09eefd8c743beb78cfd4e649e7f15e half reverted the usage
of the tests but not entirely, resulting in turning the perl dependency
which is only needed for tests into a run dependency.

The right solution would be to entirely reenable tests, (which also
depends on ruby btw) but ruby dies very quickly on the tests and the
needs to be investigated prior to renabling the test.

Bump portrevision as the running dependencies changed