devel/py-breathe: fix alternative Python versions
The port used to build only for the default python version,
use BINARY_ALIAS to help the build find the proper Sphinx version.

Switch to autoplist to permit packaging for Python 3.6/3.7, too
(__pycache__ layout differences don't work well with static pkg-plist).

Not bumping PORTREVISION again on the assumption that we would not have
been able to build a package previously on an alternative Python
version, so there is nothing to flush out with a PORTREVISION bump.