devel/py-clang: Add new port
This is a stand-alone python library binding the libclang
functionality. Currently each llvm port installs it's own version, but
this is an option (albeit default) and it's installed in a
non-standard location. This port removes the dependency on the option
and the need to patch software that uses it to use the non-standard
location.

PR:		258199