astro/py-pyerfa: New port
PyERFA is the Python wrapper for the ERFA library (Essential Routines for
Fundamental Astronomy), a C library containing key algorithms for astronomy,
which is based on the SOFA library published by the International Astronomical
Union (IAU). All C routines are wrapped as Numpy universal functions, so
that they can be called with scalar or array inputs.

WWW: https://github.com/liberfa/pyerfa