print/linux-jabref: Add new port
This is a port of JabRef. We used to have a native version of JabRef
available in the ports as print/jabref. Unfortunately, it expired last
year. print/linux-jabref uses a poratble Linux release and runs in
Linuxulator. Hopefully, this is only a temporary workaround solution for
having JabRef on FreeBSD.

PR:		257979
Requested by:	Corey Stephan
Requested by:	O. Hartmann <ohartmann@walstatt.org>