index: keep index file as long as possible
When generating the INDEX file, we are already building it in a
temporary file and swap it last minute. make the index target now
directly depend on the INDEX file generation target.
Turn that target into a PHONY target so the file is always regenerated
when someone calls make index.

PR:		202070