devel/arm-none-eabi-gcc: More flexible multilib options
This patch introduces port build options to control GCC's multilib
layout. It defaults to "disable multilib" as the port used to be set
before.

In order to build a compiler suitable for Cortex-M devices, select the
"Optimized for Cortex-M/R" option instead.

A subsequent patch to devel/arm-none-eabi-newlib will automatically
create their pkg-plist which now depends on the multilib option
chosen.

Reviewed by:	manu
Differential Revision:	<https://reviews.freebsd.org/D32638>