cad/ngspice_rework: Eliminate conflict between flavors
Before both flavors installed headers and .cm files.
This caused them to conflict with each other.

This commit removes conflicting files from the @x11 flavor
allowing the executable to be installed in parallel with libraries.

The executable seems to work in a standalone mode fine, but if it ever
needs headers or .cm files the @shlib flavor can always be installed.

PR:		259809
Tested by:	kevinz5000@gmail.com