lang/silq: Fix build failures due to llvm13 changes
The visibility rules of some variables have been changed in LLD13,
leading to link failures.

This is no issue if LDC is built with LLVM13, but the version
currently in the ports tree  is too old to be build with LLVM13.

As a work around and since the currently provided LDC is built with
LLVM10, make this port depend on LLVM10, too.

Update to the latest development source level on this occasion.