www/xcaddy: add port
This command line tool and associated Go package makes it easy to
make custom builds of the Caddy Web Server.

Custom caddy builds are required to activate 3rd party plugins.

The resulting 'caddy' executable is not part of the package, and
you will have to manage it on your own.