Mk/Uses/jpeg.mk: Update comment for graphics/mozjpeg
MozJPEG (graphics/mozjpeg) is a patched libjpeg-turbo (graphics/jpeg-turbo). It
is also a valid JPEG implementation.

from README.md:
MozJPEG is compatible with the libjpeg API and ABI. It is intended to be a
drop-in replacement for libjpeg. MozJPEG is a strict superset of libjpeg-turbo's
functionality. All MozJPEG's improvements can be disabled at run time, and in
that case it behaves exactly like libjpeg-turbo.

Reference:	https://github.com/mozilla/mozjpeg/blob/master/README.md
PR:		257028
Approved by:	bapt (portmgr, maintainer)