emulators/wine: revert accidentially committed local changes
In commit 5933ac0b099 CONFLICTS_INSTALL definitions in multiple ports
have been fixed to omit the unneeded version patterns they had.

The commit of the wine port did not include that fix, but instead
local changes to deal with an issue introduced by recent changes to
sched.h.

This commit reverts the accidentially committed changes, but retains
the intended removal of a version wildcard in CONFLICTS_INSTALL.

Reported by:	gerald