multimedia/zoneminder: remove now useless ONLY_FOR_ARCHS restriction
It was added in r486990 (f9e7f55b3a20d) without a proper explanation,
and now with seven items on this list, including all Tier-1 and most
Tier-2 architectures, it had clearly outgrown the intended purpose of
this knob.  There shouldn't really be anything architecture-specific
about this software in the first place.