sysutils/rsyslog8: update pkg-message and add sample config file
Upstream changes mean that the advice formerly given in pkg-message to
modify the default /etc/syslog.conf file as the basis of a new
${PREFIX}/etc/rsyslog.conf are now less useful and confusing to end
users.

Therefore, update the pkg-message accordingly, and install a sample
config file which should duplicated the functionality of the base
syslog.conf.

PR:		260158
Reported by:	Mark Johnston