biology/libgff: GFF/GTF parsing library based on GCLib
This is an attempt to perform a simple "libraryfication" of the GFF/GTF
parsing code that is used in GFFRead codebase. There are not many
(any?) relatively lightweight GTF/GFF parsers exposing a C++ interface,
and the goal of this library is to provide this functionality without
the necessity of drawing in a heavy-weight dependency like SeqAn. Note:
This library draws directly from the code in GFFRead and GCLib, and
exists primarily to remove functionality (and hence code) that is
unnecessary for our downstream purposes. In the future, it may be
appropriate to just replace this library wholesale with GCLib.