KDE: Update KDE Gear to 21.12
KDE Gear 21.12 has landed and comes with a massive number of updates and
new versions of applications and libraries. Literally, dozens of classic
KDE everyday tools and the specialised sophisticated apps you use to
work, be creative and play, are getting refreshers with design
improvements, new features and performance and stability enhancements.

What's new:
	https://kde.org/announcements/gear/21.12.0/#whatsnew

Changelog:
	https://kde.org/announcements/changelogs/gear/21.12.0/