mail/mailman: regression fix update to 2.1.39
Mark Sapiro announced Mailman 2.1.39 "[...] fixes
https://bugs.launchpad.net/mailman/+bug/1954694
[...]
The fix for CVE-2021-42097 was case sensitive and should not be.
The fix for CVE-2021-44227 introduced a potential NameError in logging.

This could cause a user's changes to the option's page to not be
accepted and perhaps cause a 'We hit a bug' response if the user visited
the page with a mixed- or upper-case email address."

URL:		https://bugs.launchpad.net/mailman/+bug/1954694
MFH:		2021Q4