databases/pgbarman: Update to 2.17
Following upstream changes, switch master site to GitHub, including
for hte pre-build PDF manual.

Drop support for python-3.6 -- the software claims to support this
version, but setup.py fails with an obscure error, and considering
that python-3.6 is going out of support at the end of this month, I
could not justify the time to develop a compatibility fix.

Drop the USE_PYTHON 'optsuffix' flag in favour of an explicit
PKGNAMESUFFIX

As reported in PR 260340, appending the python version to the barman
${ETCDIR} doesn't make sense.  Fix this by dropping the USE_PYTHON
'concurrent' flag, which doesn't really make sense for end-user
software rather than library modules. (1)

PR:		260340 (1)
Reported by:	bsd at abinet.ru