devel/sonarqube-community: update to 9.2.4
This update contains an updated elasticsearch which contains an
updated log4j (2.17.0).

According to Elastic this is a "feel-good" release to prevent a
false positive log4j detection of a scanner. Elasticsearch is not
vulnerable in the previously boundled version according to Elaastic,
as described in the previous commit.

PR:	260421