sysutils/slurm-wlm: Release to ports@
Newer versions of SLURM contain unconditional dependencies on cgroups.
This is a major addition to other long-standing linuxisms such as
parsing /proc files to enforce resource limits (this should be done
using sysctls on FreeBSD).

Currently, it would *not* be extraordinarily difficult to make SLURM
mostly functional on FreeBSD, but it has also become increasingly
complex and unsuitable for small-scale HPC.  Given the current
direction and the apparent priorities of SchedMD, I believe it will
only become more difficult to maintain non-Linux ports over time.

Rather than fight the current, I have begun development of a new batch
system with a commitment to simplicity and portability:

https://github.com/outpaddling/lpjs