math/py-numpy: work around build issue on -CURRENT
A sched_getaffinity function has been added to -CURRENT and the
Python build detects it presence and provides a Python function
os.sched_getaffinity().

Apparently the C library function is not correctly wrapped and
returns an out of buounds result.

Comment out the call of os.sched_getaffinity() call until this
function works correctly in the Python interpreter.

No port revision bump, since this is just a temporary fix for the
broken build on -CURRENT.

Reported by:	antoine