emulators/wine-devel: update to 7.0-rc4
The 4th milestone of the soon to be released Wine 7.0 with 38 bug
fixes, and another 22 bug fixes from the (skipped) RC3 release.

In addition to the patch provided by Damjan, part of the PLIST change
in PR 260612 has been applied to install some files that used to be
dependent on the STAGING option.