devel/rubygem-resque-scheduler: Add rubygem-resque-scheduler 4.4.0
Resque-scheduler is an extension to Resque that adds support for queueing items
in the future. Job scheduling is supported in two different ways: Recurring
(scheduled) and Delayed. Scheduled jobs are like cron jobs, recurring on a
regular basis. Delayed jobs are resque jobs that you want to run at some point
in the future.

WWW: https://github.com/resque/resque-scheduler