devel/llvm-devel: fix i386 builds, minor improvements
Fix OpenMP build on i386 by removing a bogus assert. The assert was
added in https://reviews.llvm.org/D114270 and is wrong on 32-bit. It's
a little unclear if this will work of if we should disable the AMDGPU
plugin on i386.

Remove some conflicts for long-removed packages (clang-devel and
iomp5-devel).

Don't hardcode PYTHON_EXT_SUFFIX in the plist. [0]

PR:		261034 [0]
Reported by:	Trond.Endrestol@ximalas.info [0]