net-p2p/radarr: Update to 3.2.2.5080
This release is the first to be be built with CoreCLR/.NET which is
superceding the Mono project and future versions (4.x)
of Radarr will only work with the .NET language implementation.

Previously we used binary builds of Radarr and the Mono runtime would
be used to execute the binary. The new .NET implementation doesn't
require a runtime as the compiled binary is natively executable.

Compiling Radarr with Mono was difficult in the past so we shipped the
binary and the Mono runtime, but we expect in the future we will be able
to compile Radarr and other similar apps directly with a FreeBSD
supported release of .NET.

I am also releasing myself as maintainer as we have several suitable and
motivated people keeping up with these latest developments and I don't
want to continue to block progress.

PR:	259194