mail/heirloom-mailx: Fix wrong UTC offset for Europe/Dublin
Somewhere around 2018, the tzdata maintainers (IANA) corrected a
historical mistake with the Europe/Dublin timezone. The mistake was
rooted in a misunderstanding of whether IST meant "Irish Summer Time"
or "Irish Standard Time".

The problem was discussed at great length
(http://mm.icann.org/pipermail/tz/2018-January/thread.html) and it was
concluded that IST really meant Irish *Standard* Time (in constrast with,
say, British *Summer* Time), and that this standard time is defined as
UTC+0100.

This corresponds to the article at
https://en.wikipedia.org/wiki/Time_in_the_Republic_of_Ireland and the
notes at https://en.wikipedia.org/wiki/Winter_time_(clock_lag); the
source archive of tzdata has a long section dedicated to this problem and
a large set of official references and links to www.irishstatutebook.ie.

Once the question was settled, the only possible solution for keeping
the Irish local time in sync with the rest of the world (timezones aside,
the local time in Ireland - as understood by common people - is the same
as in London and Belfast) was for IANA to _reverse_ the functioning of
the DST flag for Ireland. The result is that in the current IANA timezone
database (2021e), Europe/Dublin has DST applied in *winter*, with an
adjustment of -1h (that is, negative).

Digging deeper, one uncovers that there are a few other countries that
have (or once had) the same time-switch mechanism as Ireland; amongst
others, https://github.com/MenoData/Time4J/issues/742 also concedes that
negative DST is a reality.

In heirloom mailx, the logic that works out the UTC offset does the right
thing up to a point (November 2021, Ireland = UTC+0100), but then upon
inspecting tm->tm_isdst it sees that DST is in effect (remember, flag has
been reversed, so DST in Ireland is on in winter time) it adds one hour
(it should subtract one, because the adjustment is negative, but mailx
doesn't know).

PR:		260137
Submitted by:	Andrea Biardi <a.biardi@tiscali.it>
Reported by:	Andrea Biardi <a.biardi@tiscali.it>
MFH:		2022Q1