www/dokuwiki: Mark IGNORE with PHP 8.0 or later
Current latest release of DokuWiki isn't compatible with PHP 8.
Development is currently underway by upstream, and it seems to be
expected that new version that supports PHP 8 will be released before
the EoL of PHP 7.4. But anyway mark IGNORE with PHP 8.0 or later for
now.

PR:		261656
Approved by:	delphij (maintainer)