devel/qt5-qdoc: fix CMake Config file for doctools
We patch Qt to install into $LOCALBASE/lib/qt5, but the CMake
bits of Qt go into $LOCALBASE; this mismatch of installation-
prefixes means that the CMake Config files are generated with
an incorrect number of "../" components, and end up looking in
the wrong prefix. Symptom is this:

  The package "Qt5DocTools" references the file
     "/usr/lib/qt5/bin/qdoc"

The logic to use the correct path is there, but not selected
when writing out the Config file. Delete all the ones we don't
want, rather than letting qmake / whatever text-substitution tool
is in the build do it.

Suggested by:	tcberner