KDE: Update KDE Plasma Desktop to 5.24
Today the KDE Community releases Plasma 5.24, a Long Term Support (LTS)
release that will receive updates and bugfixes until the final Plasma 5
version, before we transition to Plasma 6.

This new Plasma release focuses on smoothing out wrinkles, evolving the
design, and improving the overall feel and usability of the environment.

Things to look for in Plasma 5.24 can be found in the anncouncement:
	https://kde.org/announcements/plasma/5/5.24.0/

The full changelog can be found here:
	https://kde.org/announcements/changelogs/plasma/5/5.23.5-5.24.0/