net/asterisk18: Fix build when textproc/xmlstarlet is installed
Asterisk build system checks for presence of xmlstarlet as "xml"
binary on the system.

If found it triggers part of the build system trying to download
some external files during the install phase which even uses a
command line tool not available on FreeBSD.

Such code is not necessary for the port, as all supported modules
are downloaded using ports tree provided functionality. Furthermore
downloading files from the internet is forbidden during the install
phase.

This patch forces the variable used by the Makefiles to identify
xmlstarlet presence to be empty. so the code path described above
is not triggered.

PR:		261884