sysutils/govmomi: Add new port
Go library for the VMware vSphere API.

govc is a vSphere CLI built on top of govmomi.
The CLI is designed to be a user friendly CLI alternative to the GUI and
well suited for automation tasks. It also acts as a test harness for the govmomi
APIs and provides working examples of how to use the APIs.

vcsim - A vCenter and ESXi API based simulator
This package implements a vSphere Web Services (SOAP) SDK endpoint intended
for testing consumers of the API. While the mock framework is written in the
Go language, it can be used by any language that can talk to the vSphere API.

WWW: https://github.com/vmware/govmomi/

PR:	261672