security/sssd: implement detection of mempcpy at configure-time
Newer versions of FreeBSD provide mempcpy, which conflicts with the port's
own implementation -- meant for the older versions.

While here, fix compilation with newer Python -- not just 3.7 -- and relax
the requirement for Python-version. This necessitated bumping PORTREVISION --
because the default Python is 3.8 nowadays.

Also, define TEST_TARGET -- and fix one of the test to compile. This needs
further work, however.

PR:	257866