www/py-weboob: Rename to www/py-woob
Project was renamed from weboob to woob.
Update to 3.0.

www/py-weboob-qt: Rename to www/py-woob-qt

Project was renamed from weboob-qt to woob-qt.
Update to latest commit. Upstream did not create a tag in git after
the project was renamed, so a new release was never made. But we
need these changes so it is compatible with base woob. This is
technically version 3.0.