dns/py-dnspython1: New port
This port is copied from dns/py-dnspython. Assign to original author.
This port is created to streamline the updated version of 2.x release
for using with other ports.

dnspython is a DNS toolkit for Python. It supports almost all record
types. It can be used for queries, zone transfers, and dynamic
updates. It supports TSIG authenticated messages and EDNS0.

dnspython provides both high and low level access to DNS. The high
level classes perform queries for data of a given name, type, and
class, and return an answer set. The low level classes allow direct
manipulation of DNS zones, messages, names, and records.

This is 1.x version of the port which supports python 2.7

WWW: http://www.dnspython.org/