japanese/ja-tex-xdvik: New port
This port is copied from print/tex-xdvik. The port consists of some
extra patches from Japanese SourceForge page and the exact purpose is
not well understood by me. The patch is also from the early 2008 and
currently not compatible with the latest version. But if someone can
make this work with the latest version(will commit soon) patches are
welcome.

From my understanding hrs@ might be the best possible person to
understand this hence assigning MAINTAINERSHIP to hrs@.