net-im/pidgin-bot-sentry: Rename pidgin-bs to pidgin-bot-sentry
The plugin was always called pidgin-bot-sentry. The fact that the
tarball is called pidgin-bs is the reason why it was originally
added under this name.

Comparing the project name at repology [1] confirms that error
nicely:

https://repology.org/project/pidgin-bot-sentry/versions
https://repology.org/project/pidgin-bs/versions

While here:

USE_LDCONFIG is not needed for .so that's installed in the pidgin
plugin directory.

Pacify portlint/portclippy, apply portfmt.