devel/gconf2: cleanup and reduce the number of dependency
Switch the ipc mecanism from Orbit2 to dbus (ORBit2 is not used
anywhere anymore).

While here remove the default services, since gconf is not used
anymore by any desktop, having the default service started to allow
configuring system wide configuration is pointless.

Drop the option for ldap, given that no desktop uses it anymore, the ability
so deal with ldap is now pointless as well

Drop the dependency on gtk only used for gconf-sanity check

This shave the dependency lits removing rust, polkit and orbit from it
while not changing the features for the few left consumers.