databases/postgresql14-server: Avoid using llvm14
When the base system compiler is clang 14, the postgresql ports attempt
to build against the devel/llvm14 port. This will fail with several
compile errors as there are a few incompatible changes in the LLVM 14 C
and C++ APIs. Until upstream postgresql adapts to these changes, limit
the PG_LLVM_VERSION to a maximum of 13.

PR:		 261976
Approved by:	maintainer timeout (2 weeks)
MFH:		2022Q1