net/ntopng,net/ndpi: Fix build due to missing symbols in certain circumstances
When building on live machines when rrdtool happens to be already
present, which often happens automatically due to the order in which
the ports system builds things locally, the ntopng port could fail
due to unreferenced symbols from glib. This was caused by ntopng
trying to unconditionally link statically to ndpi but not providing
all required static libraries references to the command line.

- Add databases/rrdtool dependency in ndpi to ensure the same
  libraries are used in both ports
- Make ntopng unconditionally use dynamic linking for ndpi too,
  avoiding the issue described above.

PR:		262756