www/py-django: Deprecate and set to expire
This port was last used as a metaport and has only Django 2.2 as the
only option knob.  The use of this port as a dependency for several
consumers has in the past also repeatedly led to some problems with
conflicting transitive dependencies.

So there is no longer any use for this metaport thus let it expire
together with www/py-django22.

PR:		261313
Reviewed by:	dbaio (python, maintainer)
Approved by:	kai (python, maintainer)
Differential Revision:	https://reviews.freebsd.org/D34453