sysutils/upower: Fix build when xsltproc is present and MANPAGES disabled
The new meson build system performs an unconditional check for
xsltproc, even when bot gtk_doc and man options are disabled, marking
it only as required if any of those options are enabled.

if xsltproc is found building manual pages and gtk_doc files is
performed unconditionally, but, when the MANPAGES option is disabled,
gtk-doc is not required and most probably missing from the system,
causing the build to fail.

Fix by performing the check for sltproc only when one of those two
options are enabled, and disabling building all docs otherwise.

PR:		262885
Fixes:		4025875f76dd78090a158a3077ca447a05722510