devel/py-azure-data-tables: New port
This is the Microsoft Azure Data Tables Client Library.

Azure Tables is a NoSQL data storage service that can be accessed from anywhere
in the world via authenticated calls using HTTP or HTTPS. Tables scales as
needed to support the amount of data inserted, and allow for the storing of
data with non-complex accessing. The Azure Tables client can be used to access
Azure Storage or Cosmos accounts.

WWW: https://pypi.org/project/azure-data-tables