net/glib-networking: update to 2.70.1 to match glib
glib-networking and glib (glib20) should be kept in sync
with major.minor versions. Update to catch up with glib.
Bump PORTREVISION on direct consumers, but not transitively.

Note that the PR has a newer patch, but that bumps glib
by a minor revision as well, to 2.72, which is tricky in itself.

Transitive consumers of libsoup and libsoup3 all build,
except for net-im/libnice, but that fails locally without this
change as well, so I'm going to choose to ignore that.

PR:		262389
Obtained from:	yzrh@noema.org