devel/glib20: add DOCS option
There is documentation possible with glib, using gtk-doc.
Add a knob and the plist bits. The documentation depends
on the tests, though: see glib issue
	https://gitlab.gnome.org/GNOME/glib/-/issues/2430

Futz a bit with the options and dependencies; TEST only needs
dbus for the build, and the docs need the tests.

PR:		258624
Reported by:	Hodong@nimfsoft.art