editors/openoffice-devel: Fix dependency issues
Fix an incorrect coinmp-related dependency that was causing spurious
rebuilds.  It was also reported to break the build of openoffice-4,
though I was unable to reproduce the problem, and it did not seem to
break the official package build.  Add another missing dependency.
Both issues were reported by stage-qa.

PR:		263238