devel/llvm14: update to 14.0.1
Note: with LLVM 14, there will now be a series of .1 - .6 releases
cut about ever two weeks rather than a single .1 release after a series
of release candidates.

Add a link for libclang.so.14 to reduce the need to modify build system
that assume the libclang version and LLVM version are the same.