lang/gcc11: Add upstream patch to fix bug of libgccjit
After switching to LTO bootstrap with commit 2f6be9fb8281, build of
editors/emacs-devel fails with segmentation fault of emacs if
NATIVECOMP option is enabled. The cause of the problem is bug of
libgccjit and it was unveiled by switching to LTO bootstrap. So add
upstream patch to fix it.

PR:		263271
Approved by:	gerald (toolchain@)