Mk/Uses: Add pytest.mk
Add USES=pytest, a standardized pytest entry point.

This knob adds dependency on pytest and defines do-test target which
calls it "the right way" (respecting TEST_ENV, with proper verbosity,
additional info on skipped tests, and resetting addopts which may
be set by upstream and introduce unwanted options). This allows to
remove duplicate (and in many cases not completely correct) do-test
targets from a lot of pytest consumer ports.

Additionally, it adds PYTEST_IGNORED_TESTS (to list tests which
fail and are expected to fail) and PYTEST_BROKEN_TESTS (to list
tests which fail and needs to be fixed). Both knobs lead to
listed tests being skipped, which is believed to be better than
always failing `test` target, allowing to monitor status of other
tests and better document known failures.

Possible room for improvement:
- Support for pytest plugins (add dependencies and disallow pytest
  from using unlisted plugins which often cause breakages)
- Treat PYTEST_BROKEN_TESTS as xfail to catch the cases when tests
  are fixed.

  Both items seem to require support on pytest side though.

Reviewed by:		tcberner
Differential Revision:	https://reviews.freebsd.org/D34693