mail/procmail-bgrb: New port
mail/procmail-bgrb is a new port tracking the BuGlessRB fork of
procmail (which hasn't been updated for over 20 years).
The commits to BuGlessRB/procmail appear be by the original author
of procmail. Wikipedia documents that the upstream for this port is
its repo.

As this upstream is in active development and the original procmail on
sourceforge is defunct, we may wish to consider merging or replacing
mail/procmail with this port at some point.

Reported by:	Lucas Holt at midnightbsd.org (on Twitter)