devel/p5-Parse-Distname: Add p5-Parse-Distname 0.05
Parse::Distname is yet another distribution name parser. It works almost the
same as CPAN::DistnameInfo, but Parse::Distname takes a different approach. It
tries to extract a version part of a distribution and treat the rest as a
distribution name, contrary to CPAN::DistnameInfo which tries to define a name
part and treat the rest as a version.

Because of this difference, when Parse::Distname parses a weird distribution
name such as "AUTHOR/v1.0.tar.gz", it says the name is empty and the version is
"v1.0", while CPAN::DistnameInfo says the name is "v" and the version is "1.0".
See test files in this distribution if you need more details. As of this
writing, Parse::Distname returns a different result for about 200+ distributions
among about 320000 BackPan distributions.

WWW: https://metacpan.org/dist/Parse-Distname