security/sniffglue: Add new port
sniffglue is a network sniffer written in rust. Network packets are parsed
concurrently using a thread pool to utilize all cpu cores. Project goals are
that you can run sniffglue securely on untrusted networks and that it must not
crash when processing packets. The output should be as useful as possible by
default.

PR:		262959