devel/hare: rename package to not conflict with sysutils/hare
The package name has been changed to hare-lang in order to not collide
with the existing hare package.

The name of the binary ("hare") is unchanged, since the documentation
and possible support tools expect it to have that name. There is no
conflict with the binary installed by the sysutils/hare port, since
that is installed into ${PREFIXE}/sbin, while this port creates it in
${PREFIX}/bin.

If both the hare and hare-lang packages are installed, ${PREFIX}/bin
should be to the left of ${PREFIX}/sbin in the user's ${PATH}, since
${PREFIX}/sbin/hare is generally invoked using the full path from a
PAM configuration file.

Reported by:	antoine