Mk/Uses/go.mk: Do not insist on always pulling go.mod and distfile from GOPROXY
Some Go packages are not listed on pkg.go.dev (usually because their
versioning scheme does not satisfy Go requirements) and, as a
consequence, are not accessible through proxy.golang.org.

Instead of resorting to hacks with pseudo-versions and/or enumerating
dependencies manually with modules2tuple, support fetching go.mod and
distfile from the usual places using DISTFILES mechanism. This will
allow bypassing GOPROXY for go.mod and the distribution archive, but
will still use it for fetching dependencies (with GO_MODULE).