USES=cmake: be explicit about incompatibilities with ninja
Document how `USES=gmake` and `USES=emacs` are incompatible with
the default `USES=cmake` behavior of using ninja as generator.
Using a naive `USES=cmake emacs` line used to get you a strange
ninja error when ${MAKE_ARGS} is passed to it, now there is a
meaningful error message beforehand.

The linked PR examines ways to make it "just work", but that
got bogged down. Do a partial fix so it is easier to track
down next time.

PR:		251191