Mk/Uses/bdb.mk: Allow bdb:18 to work by default.
mail/mutt currently has USES=bdb:18 which causes an obscure error in
Poudriere causing it to be IGNORED unless WITH_BDB6_PERMITTED is set.
We already have a license framework so this flag seems odd but in the
case where a port explicitly is asking for the version it makes sense
to permit it as the default.