emulators/virtualbox-ose-kmod*: Tie to a specific release branch.
Use a more restricted range of supported __FreeBSD_version values for
the virtualbox kernel module since it uses internal KPIs for which KBI
stability is not guaranteed.  DECLARE_MODULE_TIED would tie the module
to a specific __FreeBSD_version which would cause the module to stop
loading when an SA or EN is pushed on a release branch.  Instead, use
DECLARE_MODULE_MAXVER with an upper bound that is limited to the range
of a releng branch so that a module built on X.Y will load on X.Y with
SAs or ENs applied but will fail to load on X.Y-1 or X.Y+1.

Reviewed by:	kevans, imp, madpilot, lwhsu (earlier version)
Differential Revision:	https://reviews.freebsd.org/D16119