devel/py-configupdater: Add py-configupdater 3.1
The sole purpose of ConfigUpdater is to easily update an INI config file with no
changes to the original file except the intended ones. This means comments, the
ordering of sections and key/value-pairs as wells as their cases are kept as in
the original file. Thus ConfigUpdater provides complementary functionality to
Python's ConfigParser which is primarily meant for reading config files and
writing new ones.

WWW: https://github.com/pyscaffold/configupdater