textproc/py-rich-cli: Add py-rich-cli 1.7.0
Rich-cli is a command line toolbox for fancy output in the terminal, built with
Rich.

Use the rich command to highlight a variety of file types in the terminal, with
specialized rendering for Markdown and JSON files. Additionally you can markup
and format text from the command line.

WWW: https://github.com/Textualize/rich-cli