www/joomla4: rename the package from joomla to joomla4
The existing www/joomla3 port uses joomla3 as package base name, but
the new www/joomla4 port has been created without a PKGNAMESUFFIX and
thus with a PKGBASE value of "joomla".

It is good practice to include the major version in PKGBASE for ports
for which more than 1 major version is supported at a time, even in
cases where one of those major versions is clearly preferred for new
installations of the software.

Approved by:	wen (maintainer)