lang/ruby27: Mark DEPRECATED and set EXPIRATION_DATE
After 2.7.6 is released on April 12, 2022, Ruby 2.7 entered the
security only maintainance phase. It continues for a year and after
that Ruby 2.7 reaches its EoL. On the other hand default version of
Ruby has switched to 3.0. So mark this port as DEPRECATED and set
EXPIRATION_DATE to the date of EoL in order to encourage user to
migrate to newer version.

PR:		263451
Approved by:	maintainer timeout