Uses/pyqt.mk: repair PORTEPOCH of py-sip dependency
When using BUILD_DEPENDS with a dependency-version, as described
in section 5.9.9 of the Porters Handbook, a desired-version can be
compared with an installed-version, e.g.:

	py38-sip>=6.5.1:devel/py-sip@py38

The version is compared fully with all the ports-versioning
parts included. That includes the PORTEPOCH, which isn't
visible (because =0 in the specification above). Since py-sip
was epoched at some point, the installed version of py-sip might
be this:

	py38-sip-5.5.0,1               Python to C and C++ bindings generator

Because of the epoch, the version 5.5.0,1 > 6.5.1, even though
visually it seems like it wouldn't be. Massage the dependency generation
in pyqt.mk to take that epoch into account.

PR:		262129
Reported by:	Rafael Grether