net/mpich2: remove
MPICH2 has been marked as deprecated since dec. 2020, and no port in the
tree depends on it anymore.

Upstream switched to MPICH3 (net/mpich) and MPICH4 (not yet in the ports
tree), and now it does not build with GCC-11: it’s time to remove it.

PR:		264366
Reported by:	salvadore