editors/vim: Update to 8.2.5052 and fix vimrc clobbering
Vim runtime's defaults.vim will clobber global vimrc settings. It's
fine for the settings we provide, but other edits to that file can
get lost.

To work around this, defaults.vim is now directly sourced and a flag
is set to stop that file from getting loaded a second time. Thanks go
to Anton Saietskii for that bit of magic.

PR:	251420