graphics/libglvnd: fix on armv6/armv7
The .thumb_func directive does not switch to thumb mode in clang's
integrated assembler.  This causes the code to be miscompiled in that
it is generated as ARM code but the symbols are tagged as thumb.  The
resulting library is unusable.  Every call into it crashes.

Fix this issue by applying an upstream patch set to have the entry
stubs be in ARM mode instead of Thumb mode.

See also:	https://gitlab.freedesktop.org/glvnd/libglvnd/-/merge_requests/267
See also:	https://gitlab.freedesktop.org/glvnd/libglvnd/-/issues/232
PR:		263845