devel/py-jsonschema: Include schemas
After the latest upgrade to 4.6.0 the pkg was missing the schemas file
as the new hatch build system which doesn't create the full files as our
building procedure requires dependency on a stub setup.py file.

For future reference:
The method I used for obtaining the setup.py file is to add

[tool.hatch.build.targets.sdist]
support-legacy = true

to pyproject.toml. Afterwards, building an sdist with `python3 -m build
-s` yields a corresponding archive in dist/jsonschema-4.6.0.tar.gz, from
which the setup.py file can be extracted. The latter build still doesn't
contain the subdirectory in MANIFEST.in though, so it didn't fix the
issue entirely.

While there are lots of dependent ports for now I am committing this but
will look into a better solution to adapt the situation.

PR:		264452
Reported by:	stefan.haemmerl@gmx.de