devel/android-tools: New port: Android debugging tools
While I'm here, remove the old devel/android-tools-adb and
devel/android-tools-fastboot ports. These ports are long out-of-date
and the adb port especially had issues with certain devices.

This port was tested from a FreeBSD 14.0 system to flash both
LineageOS and GrapheneOS on a Google Pixel 3 using both `adb` and
`fastboot`.