textproc/py-mike: Add py-mike 1.1.2
mike is a Python utility that makes it easy to deploy multiple versions of your
MkDocs-powered docs to a Git branch, suitable for hosting on Github via
gh-pages.

mike is built around the idea that once you've generated your docs for a
particular version, you should never need to touch that version again. This
means you never have to worry about breaking changes in MkDocs, since your old
docs (built with an old version of MkDocs) are already generated and sitting in
your gh-pages branch.

While mike is flexible, it's optimized around putting your docs in a
<major>.<minor> directory, with optional aliases (e.g. latest or dev) to
particularly notable versions. This makes it easy to make permalinks to whatever
version of the documentation you want to direct people to.

WWW: https://github.com/jimporter/mike