devel/py-pathos: Add py-pathos 0.2.8
pathos is a framework for heterogeneous computing. It provides a consistent
high-level interface for configuring and launching parallel computations across
heterogeneous resources. pathos provides configurable launchers for parallel and
distributed computing, where each launcher contains the syntactic logic to
configure and launch jobs in an execution environment. Examples of launchers
that plug into pathos are: a queue-less MPI-based launcher (in pyina), a
ssh-based launcher (in pathos), and a multi-process launcher (in multiprocess).

WWW: https://github.com/uqfoundation/pathos