math/vtk9: fix build after d1e8146547 (netcdf 4.9.0)
NC_SZIP_NN is not #define'd in netcdf.h, so it cannot be the target
of an assignment anymore. Fortunately NC_SZIP_NN is now defined to
the same value (32) as it was set to by vtk9 before (and it was
declared const int before), so just hide the old const int if NC_SZIP_NN
is already defined.
No functional change, no bumps.