devel/courier-unicode: remove bogus CONFLICTS entry
Due to CONFLICTS=libunicode this port could not be built or installed
on system that already had libunicode installed. Since both of these
ports are dependencies of other ports, only ports depending on one or
the other could be installed at a time.

I have verified that there is no build or install conflict, and the
libunicode port is not marked as conflicting with this port.

In order to allow ports depending on courier-unicode and libunicode to
be installed at the same time, I'm removing the bogus CONFLICTS entry.

Approved by:	portmgr (implicit)