net/gnome-connections: new port
View and use other desktops

Connections allows you to connect to and use other desktops. This can be a great
way to access content or software on a different desktop operating system. It
can also be used as a way to provide support to users who might need help.

A range of different operating systems can be connected to, including Linux and
Windows desktops. You can also connect to virtual machines.

Connections uses the widely supported VNC and RDP protocols, and one of these
must be enabled on the desktop that you want to connect to.

WWW: https://apps.gnome.org/app/org.gnome.Connections/