games/xnethack: fix behaviour with mismatched lua version
The latest version of the package has introduced a "lua sandbox" that
comes with a version check against the bundled lua version.

As we unbundle lua, we cannot ensure that this version check succeeds.
Removal of the version check allows us to build a working binary, but
it is uncertain if there are any residual effects from the version
mismatch.

To rule out that possibility, the sandbox code is left disabled by
default but can be enabled if desired through the SANDBOX option.

PR:		264882