audio/espeak: update to version 1.15.1
This port updates the espeak port to the sources distributed as
espeak-ng. The old project had been stalled and a new developer
team created a repository that continues where the previous team
stopped.

This port installs files with a suffix of -ng (e.g. espeak-ng),
but provides symbolic links that allow the binary and library to
be accessed under the legacy name without this suffix.

I could have created an audio/espeak-ng port, but since espeak-ng
should be a fully compatible replacement of espeak, I have decided
to re-use the existing port name. (The existing audio/espeak port
had been marked as deprecated and to be deleted on 2022-06-30.)

I have assigned the port to me (since I consider it to be a new
port under the name of the deprecated port), but do not object to
the previous maintainer of the audio/espeak port assigning it back
to himself.