audio/espeak-ng: add espeak to CONFLICTS_INSTALL
The espeak package has been replaced by espeak-ng, but since the pkg
command does not understand that the latter replaces the former, it is
possible that "pkg upgrade" tries to install espeak-ng on a system
that already has espeak installed.

Create and plan to keep the CONFLICTS_INSTALL entry for some time in
order to let the pkg command know that it cannot install espeak-ng
without prior deinstallation of espeak.