graphics/svgalib: prevent "make all" from deleting files
The Makefile provided with this software made "make install" imply
"make uninstall" as a pre-condition. This caused the installation of
files to the stage directory to be preceded by the removal of those
files from the base system.

This is unacceptable behavior for a port: no files outside of the
work directory and possibly $TMPDIR (i.e. /tmp) must be affected by
"make all" to build a port.

This issue did not affect building a package in poudriere, since
there were no installed files to delete, and the generated package
was correct. Therefore, the PORTREVISION has not been bumped.

Reported by:	Andy Farkas <andyf@andyit.com.au>