sysutils/ansible: Do not replace shebangs within modules
It turns out that Ansible does not like when we replace the shebang
lines with a path to a Python interpreter with a trailing number
(like "python3.9").

Stop replacing the shebang lines in modules directories. If it turns out
to be problematic as well, we may consider introducing an indirect
symlink to the versioned Python interpreter from a path like
/usr/local/libexec/ansible/bin/python for example.

PR:		265211
Reported by:	vladimir-csp@yandex.ru