emulators/wine: Fix crash due to ASLR
ASLR has been enabled by default on -CURRENT and -STABLE and this
causes Wine to crash. The clean solution would be to implement some
memory mapping code that already exists for macOS and Linux, but this
is not an essy task.

In order to allow use of Wine on systems with ASLR mark the Wine
executables as not compatible with ASLR using elfctl. This allows
to keeps ASLR enabled on the system for all other binaries.

If the required memory mapping functionality is made available, the
elfctl commands should be removed from this and the other Wine ports.

Approved by:	portmgr (implicit)