devel/avalon-framework: Remove deprecation notice
The deprecation notice was added by mistake.

avalon-framework is used by math/scilab. SciLab is a major
software package for systems modeling, visualization, numeric
analysis, etc.

SciLab is used in multiple industries, and I am sure that
it has FreeBSD users. GUI is its main feature and it can't be
deprecated.

Its dependency (devel/avalon-framework) can't be deprecated
just because it was abandoned upstream. It is buildable and
runnable, and should stay for the time being.

Approved by:	portmgr (blanket; correct mistake)