shells/zsh: remove the modification of the configure script
We used to modify the configure script to avoid zsh to link against
librt because clock_gettime on FreeBSD is in libc not librt.

It's been a while that we are configuring zsh with the --as-needed
linker flag which means now even if -lrt is passed we will not end
up linked to it librt anymore.

besides, modifying the configure script at the patch phase makes it
newer than stamp.h.in which results in the build system triggering a
regeneration of itself (hence the USES= autoreconf we had), not
touching anymore that script can remove the need for USES=autoreconf
and the load if build dependencies it drags in

Removing the USES=autoreconf highlights also the lack of build
dependency on perl which is needed to be able to the help files for
 run-help to work on builtins. this is fixed now