databases/redis{6,62}: Fix problem that pkg-message is misleading
When installing package of databases/redis62, following pkg-message is
displayed.

          To setup "redis62" you need to edit the configuration file:
          /usr/local/etc/redis62.conf

          To run redis from startup, add redis62_enable="YES"
          in your /etc/rc.conf.

But path of configuration file is /usr/local/etc/redis.conf and what
should really be added to /etc/rc.conf is redis_enable="YES". Same
problem also happens when installing databases/redis6

The source of the problem is that the value of ${PORTNAME} is used to
create pkg-message from file/pkg-message.in. So fix it by changing the
value of ${PORTNAME} to "redis" and set PKGNAMESUFFIX in Makefile of
each port.

While I'm here,

* Pet portlint and portclippy
* Tidy up Makefile with portfmt.

PR:		265623
Reported by:	FiLiS
MFH:		2022Q3