java/openjdk11: revert accidental commit 3d803e446071
Revert part of commit 3d803e446071 that should only have affected the
deskutils/calendar port.

This patch that has been committed by accident had already been
attached to PR 265687 to fix the build issue of openjdk11 on -CURRENT,
but it is up to the java@ team to apply this or some other fix.

Approved by:	portmgr (implicit)