devel/llvm15: Add LLVM 15
Modify the llvm14-derived build for 15.0.0rc1, the first LLVM 15
release candidate.

Disble OPENMP by default on i386. It's clearly not being tested on
32-bit upstream.

Don't enable MLIR by default. Instead, rely on FLANG's dependency
on it to turn it on for architectures that support FLANG. This removes
MLIR from default packages on 32-bit system.