textproc/groff: allow to build preconv(1) without encoding autodetection
After commit f515b0e657e7, I've received another complaint that, quote,
"the number of dependencies of [this] groff thing is quite unreasonable"
and entails installing e.g. Ninja and Meson, which in turn pulls Python,
etc.  This happens because `textproc/uchardet' which can be used by the
preconv(1) utility for automatic encoding detection is built with CMake.

As this dependency is optional upstream, allow conservative ports users
to disable it and thus avoid pulling CMake and all that stuff.