audio/musescore: ignore qmake (the Qt3 one)
If devel/qmake (the Qt3 version) is installed, it is un-suffixed
and is found as "qmake", rather than the current "qmake-qt5" which
is what Musescore needs.

Patch out the name "qmake" entirely.

Submitted by Mamoru Iwaki in the linked PR. I don't know if we have
a policy on adding people's email addresses out of the PR database,
so I left it out in this commit.

PR:		265775